feat: implement a local yq check along with a docker check
In CI there is no docker(-in-docker) but a local yq is available. So this implementation checks if
there is a local yq and is at correct version (3) if so, use local execution, if yq is not installed
then use docker and the docker image to run it. If neither is installed, exit en throw error

#268