refactor: move has_docker checks to common.sh and use it to check docker and run yq in container
Created a shadow function yq, instead of dyq. Move a couple of variables from otomi to common.sh
that are necessary in both common & otomi. Also added `function` keyword for every function that
misses it, to bring it in line to the same style as otomi & bootstrap. Also fix bug in otomi.sh
`check_docker_running` where the exit code was used, now it uses `has_docker` that uses a similar
check, but is now in common.sh

fix #271