fix(values/): required changes to prevent 'pods cannot be evicted' for e.g. node reboots (kured)
When Kured is installed it needs to be able to drain nodes, this is only possible if all pdbs have
at least 1 allowed disruption. PDBS have not been changed, but minimum number of replicas have. For
knative this is a temporary fix until another solution is offered where <5 replicas is possible. See
https://github.com/knative/operator/issues/376