fix(policies): fix allowedusers policy
current annotation that is set by istio mutation will prevent complete pod to be checked on the
provided policies (allowedusers,capabilities and containerlimits). With this change it will only
skip validation for istio-init container

BREAKING CHANGE: This will cause the policies allowedusers/capabilities/containerlimits to work on
all pods that are injected by istio. This might prevent users for instance to update pods that are
running as root already.